Collections: move filter toolbar to a new CollectionFilter component
